Added new generator for lorem ipsum string.
New 'generate_iplum' generator will return a generic 'Lorem ipsum'
string. The available arguments are: words and paragraphs.

When no arguments are used, a single paragraph containing the original
'Lorem ipsum' string will be returned.

When 'words' is used, a string with the chosen number of words will
be returned, repeating the same text if necessary.

If 'paragraphs' is used, then 1 or more paragraphs of the same length
and content will be returned.

If both 'words' and 'paragraphs' arguments are used, then 1 or more
paragraphs of exactly 'words'-length will be returned, each paragraph a
continuation of the previous one, repeating the text if necessary.