Polish documentation
Fix a typo in the readme.

Fix errors in the API documentation. Remove all references to `obj` and
`object`. An "obj" cannot be instantiated, telling the user that an "object" is
returned is nearly useless (what _isn't_ an object?), and neither is listed in
`nitpick_ignore` in `docs/conf.py`. Also: Add a link to RFC 3629, drop some
spurious backslashes and better wrap a line.