Add __getattr__ tests and fix bug
Previous __getattr__ implementation was making use of locals which was
not returning the proper object list as locals in that scope was
returning only objects from the function's scope.

Instead of using globals create a "private" variable to hold all the
factories and make __getattr__ implementation easier.

Also add tests and make sure that __getattr__ implementation returns the
expected objects.